Soporte inicial para migración y creación de esquema BD
Se refactoriza `DatabaseInitializer` para depender de la nueva interfaz `IDatabaseSchemaRepository`, eliminando el acceso directo a la tabla `App`. Se implementa `DatabaseSchemaRepository` para gestionar la creación y actualización del esquema de base de datos mediante scripts SQL embebidos, incluyendo el script `database-v1.sql` que crea todas las tablas e índices principales. Se actualiza el registro de dependencias y el tipo de datos en `AppRecord` de `int` a `long` para mayor capacidad. Cambios alineados con Clean Architecture, DDD y SOLID.

+++ RememberMePlusApp/Infrastructure/Data/AppRecord.cs
@@ -1,7 +1,7 @@
 namespace RememberMePlusApp.Infrastructure.Data;
 
 public sealed record AppRecord(
-    int IdApp,
-    int Version,
-    int RelativeOffsetMinutes,
-    int SnoozeMinutes);
+    long IdApp,
+    long Version,
+    long RelativeOffsetMinutes,
+    long SnoozeMinutes);
